antimicrobial-usage-alerts/requirements.txt
# Core dependencies
requests>=2.28.0
python-dotenv>=1.0.0

# Epic FHIR OAuth (optional, for production)
PyJWT>=2.8.0
cryptography>=41.0.0
